requirements.txt
discord
django
python-dotenv
asgiref